requirements.txt
torch==2.6.0
tqdm==4.67.1
Pillow==11.1.0
numpy==2.2.3
tensorboardX==2.2
matplotlib==3.10.1
thop==0.1.1.post2209072238